app/requirements.txt
aiofiles==0.5.0
aiosqlite==0.17.0
alembic==1.7.3
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.24.0
backcall==0.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.4
cryptography==3.4.8
databases==0.5.2
decorator==5.1.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
environs==9.3.3
fastapi==0.68.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
install==1.3.4
ipython==7.28.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
matplotlib-inline==0.1.3
orjson==3.6.3
parso==0.8.2
passlib==1.7.4
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.20
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
PyJWT==2.1.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.4.25
starlette==0.14.2
traitlets==5.1.0
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.13.4
watchgod==0.7
wcwidth==0.2.5
websockets==8.1
